+++ CHANGELOG.md
@@ -11,7 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Issue #130: Add weights constraint
 - Issue #137: Robot executor
-- 
+- Issue #139: Imu failure report
+
 ## [0.4.1] 2022-11-26
 
 ### Added